Learn CSS Variables Build City Skyline - Step 07
You can see the body (it's the inner-most box on your page); the box around it is the html element. Make your body fill the whole viewport by giving it a height of 100vh. Remove the default margin from the body by setting the margin to 0.
Finally, set the overflow property to hidden to hide any scroll bars that appear when something extends past the viewport.