Learn CSS Variables Build City Skyline - Step 21
The buildings are too spaced out. Squeeze them together by adding two empty div elements to the top of the .background-buildings element, two more at the bottom of it, and one more in between .bb3 and .bb4.
These will be added as evenly-spaced elements across the container, effectively moving the buildings closer to the center.