Learn CSS Variables Build City Skyline - Step 25
That didn't work, because the variables you declared in .bb1 do not cascade to the .bb2 and .bb3 sibling elements. That's just how CSS works. Because of this, variables are often declared in the :root selector. This is the highest level selector in CSS; putting your variables there will make them usable everywhere.
 Add the :root selector to the top of your stylesheet, and move all your variable declarations there.