Learn CSS Variables Build City Skyline - Step 29
You want the .foreground-buildings container to sit directly on top of the .background-buildings element. Give it a width and height of 100%, set the position to absolute, and the top to 0.
This will make it the same size as the body and move the start of it to the top left corner.