Learn CSS Variables Build City Skyline - Step 115
Copy and paste your whole sky class along with all of its properties and values into the media query. You are going to make another color scheme for the skyline that changes it from day to night.

Note: You are going to need to scroll past the editable region to copy the class.