Learn CSS Grid by Building a Magazine - Step 07
After your header element, create a div with the class set to author.

Within that div, create a p element with the class set to author-name and give it the text By freeCodeCamp. Wrap the freeCodeCamp portion in an a element with the href set to https://freecodecamp.org, and the target set to _blank.

Below that, add a second p element with the class publish-date and the text March 7, 2019.