Learn CSS Grid by Building a Magazine - Step 12
Create another p element below your .first-paragraph element, and give it the following text:

After years - years - of pondering these two problems and how to solve them, I slipped, hit my head on the sink, and when I came to I had a revelation! A vision! A picture in my head! A picture of this! This is what makes time travel possible: the flux capacitor!