Learn CSS Grid by Building a Magazine - Step 27
Create an html selector and give it a font-size property set to 62.5%. This will set the default font size for your web page to 10px (the browser default is 16px).

This will make it easier for you to work with rem units later, as 2rem would be 20px.

Also, set the box-sizing property to border-box.