Learn CSS Grid by Building a Magazine - Step 52
If you wanted to add more social icons, but keep them on the same row,
you would need to update grid-template-columns to create additional columns.
As an alternative, you can use the grid-auto-flow property.

This property takes either row or column as the first value,
with an optional second value of dense. grid-auto-flow uses an auto-placement algorithm to adjust the grid layout. Setting it to column will tell the algorithm to create new columns for content as needed. The dense value allows the algorithm to backtrack and fill holes in the grid with smaller items, which can result in items appearing out of order.

For your .social-icons selector, set the grid-auto-flow property to column.