Learn CSS Grid by Building a Magazine - Step 64
A quote is not really a quote without proper quotation marks. You can add these with CSS pseudo selectors.

Create a .quote::before selector and set the content property to " with a space following it.

Also, create a .quote::after selector and set the content property to " with a space preceding it.