Learn build CSS Animation Ferris Wheel - Step 24
To make your cabin animation seem more like a natural swinging motion, you can use the ease-in-out timing function. This setting will tell the animation to start and end at a slower pace, but move more quickly in the middle of the cycle.

Replace linear to ease-in-out in the .cabin selector.